requirements.txt
streamlit~=1.47.0
python-dotenv~=1.1.0
cohere~=4.45.0
fpdf~=1.7.2
git+https://github.com/openai/whisper.git
